+++ ChangeLog.md
@@ -2,6 +2,7 @@
 ## 2.11
 ### Capacités
 * Commandant de la voie du chef d'armée.
+
 ## 2.10
 ### Capacités
 * Chair à canon de la voie du PNJ récurrent