+++ ChangeLog.md
@@ -3,6 +3,9 @@
 ### Capacités
 * Commandant de la voie du chef d'armée.
 
+### Corrections de bugs
+* Prise en compte de la taille des cellules de la grille des cartes, car elle est utilisée en jeu pour mesurer les distances.
+
 ## 2.10
 ### Capacités
 * Chair à canon de la voie du PNJ récurrent